refactor: split admin module into focused sub-modules
The convex/admin.ts file had grown to 1585 lines and 30+ exports spanning
8 unrelated responsibilities: auth, user management, voucher management,
analytics, feedback, health checks, image cleanup, and OCR eval pass-throughs.
Its interface was nearly as complex as its implementation — callers wanting
to expire a voucher had to navigate the same file that handled login
sessions, health checks, and image cleanup.

Split into 7 focused sub-modules under convex/admin/:

- admin/auth: login, logout, checkSession, verifyAdminSession,
  adminQuery/Mutation/Action wrappers, cleanupExpiredSessions
- admin/users: ban, flag, unban, getAllUsers, getBannedUsers, getFlaggedUsers,
  getUsersWithStats, getUserDetails, backfillUserStats, clearUserData
- admin/vouchers: getTodaysVouchers, getAllVouchers, expireVoucherAndDeductCoins,
  reverseClaim, clearReportAndUpdateVoucher, sendMessageToUser
- admin/healthChecks: performHealthCheck, runHealthCheck, runOcrEvals,
  runSingleOcrEval (OCR evals fold into health checks as system validation)
- admin/imageCleanup: getExpiredVouchersForCleanup, markVoucherImagesForDeletion,
  deleteVoucherImages, cleanupExpiredVoucherImages
- admin/analytics: getMessageAnalytics, getAnalyticsEventCounts,
  getTransactionTotalsByType, getUserGrowth
- admin/feedback: getAllFeedback, updateFeedbackStatus, getFailedUploads,
  getSampleVoucherImageUrl

Each sub-module is self-contained: queries the database directly, and
cross-module calls go through the Convex internal API (e.g.
internal.admin.vouchers.getTodaysVouchers).

Frontend and test imports updated to use the new namespaced paths
(api.admin.users.banUser, internal.admin.imageCleanup.*, etc.). The old
admin.ts is deleted — it passes the deletion test cleanly.

All 81 backend tests pass. No new type errors introduced.
